fix: rota da bíblia dinâmica
